Move widget code into cubed diagnostics
