+++ cubed/array_api/array_object.py
@@ -22,8 +22,8 @@ from cubed.array_api.linear_algebra_functions import matmul
 from cubed.backend_array_api import namespace as nxp
 from cubed.core.array import CoreArray
 from cubed.core.ops import elemwise
+from cubed.diagnostics.widgets import get_template
 from cubed.utils import itemsize, memory_repr
-from cubed.vendor.dask.widgets import get_template
 
 ARRAY_SVG_SIZE = (
     120  # cubed doesn't have a config module like dask does so hard-code this for now
