+++ cubed/diagnostics/widgets/core.py
@@ -1,27 +1,15 @@
-import datetime
 import html
 import os.path
 
 from jinja2 import Environment, FileSystemLoader, Template
 from jinja2.exceptions import TemplateNotFound
 
-from cubed.vendor.dask.utils import (
-    format_bytes,
-    format_time,
-    format_time_ago,
-    key_split,
-    typename,
-)
+from cubed.utils import format_int, memory_repr
 
 FILTERS = {
-    "datetime_from_timestamp": datetime.datetime.fromtimestamp,
-    "format_bytes": format_bytes,
-    "format_time": format_time,
-    "format_time_ago": format_time_ago,
+    "format_int": format_int,
     "html_escape": html.escape,
-    "key_split": key_split,
-    "type": type,
-    "typename": typename,
+    "memory_repr": memory_repr,
 }
 
 TEMPLATE_PATHS = [os.path.join(os.path.dirname(os.path.abspath(__file__)), "templates")]
@@ -31,7 +19,6 @@ def get_environment() -> Environment:
     loader = FileSystemLoader(TEMPLATE_PATHS)
     environment = Environment(loader=loader)
     environment.filters.update(FILTERS)
-
     return environment
 
 
@@ -39,6 +26,4 @@ def get_template(name: str) -> Template:
     try:
         return get_environment().get_template(name)
     except TemplateNotFound as e:
-        raise TemplateNotFound(
-            f"Unable to find {name} in dask.widgets.TEMPLATE_PATHS {TEMPLATE_PATHS}"
-        ) from e
+        raise TemplateNotFound(f"Unable to find {name} in {TEMPLATE_PATHS}") from e
