+++ cubed/diagnostics/widgets/templates/array.html.j2
